Safari-611.1.10/LayoutTests/accessibility/mac/update-children-when-aria-role-changes-expected.txt

button 1
This tests that when an ARIA role changes, the AX hierarchy is updated accordingly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parent.childAtIndex(0).role is 'AXRole: AXButton'
PASS parent.childAtIndex(1) == null || parent.childAtIndex(1).isValid == false is true
PASS parent.childAtIndex(0).role is 'AXRole: AXStaticText'
PASS parent.childAtIndex(1).role is 'AXRole: AXButton'
PASS successfullyParsed is true

TEST COMPLETE
